Reorganizando e restruturando dados e comunicação do menu
